feature(ui-ux): add new defichain logo
